move tree data related things from row grouping to tree data module
