+++ packages/ag-grid-community/src/interfaces/iClientSideNodeManager.ts
@@ -1,4 +1,5 @@
 import type { RowNode } from '../entities/rowNode';
+import { StageExecuteParams } from './iRowNodeStage';
 import type { RowDataTransaction } from './rowDataTransaction';
 import type { RowNodeTransaction } from './rowNodeTransaction';
 
@@ -30,4 +31,7 @@ export interface IClientSideNodeManager<TData = any> {
     updateRowData(rowDataTran: RowDataTransaction<TData>): ClientSideNodeManagerUpdateRowDataResult<TData>;
 
     setMasterForAllRows?(rowNodes: RowNode<TData>[] | null | undefined, shouldSetExpanded: boolean): void;
+
+    // TODO: this has to be removed in next PRs
+    executeTreeStage?(params: StageExecuteParams): void;
 }
