+++ packages/ag-grid-enterprise/src/rowGrouping/groupStage/batchRemover.ts
