+++ packages/ag-grid-enterprise/src/treeData/treeStrategy/treeNode.ts
